Rename and move remoteBranches.ts -> gitBranch.ts
