+++ apps/desktop/src/routes/[projectId]/+layout.ts
@@ -5,6 +5,7 @@ import { BaseBranchService } from '$lib/baseBranch/baseBranchService';
 import { CloudBranchCreationService } from '$lib/branch/cloudBranchCreationService';
 import { BranchListingService } from '$lib/branches/branchListing';
 import { BranchDragActionsFactory } from '$lib/branches/dragActions.js';
+import { GitBranchService } from '$lib/branches/gitBranch';
 import { CommitDragActionsFactory } from '$lib/commits/dragActions.js';
 import { CommitService } from '$lib/commits/service';
 import { StackingReorderDropzoneManagerFactory } from '$lib/dragging/stackingReorderDropzoneManager';
@@ -13,7 +14,6 @@ import { HistoryService } from '$lib/history/history';
 import { SyncedSnapshotService } from '$lib/history/syncedSnapshotService';
 import { ProjectMetrics } from '$lib/metrics/projectMetrics';
 import { ModeService } from '$lib/modes/service';
-import { GitBranchService } from '$lib/stores/remoteBranches';
 import { UncommitedFilesWatcher } from '$lib/uncommitedFiles/watcher';
 import { BranchController } from '$lib/vbranches/branchController';
 import { UpstreamIntegrationService } from '$lib/vbranches/upstreamIntegrationService';
